<feature> : <testerInfoPage, TesterItem 분리>
굳이 props 속성이 필요없는 컴포넌트들 간의 관계를 분리했다.
프로필 사진을 호출하기 위한 get 요청 url을 만들고 img 태그에 적용했다.

Resolves: #15
See also: #issueNo
