+++ src/views/tester/TesterInfoPage.vue
@@ -1,11 +1,7 @@
 <template>
   <div>
     <h2>Tester Info Page</h2>
-    <TesterItem
-      :key="testerItem._id"
-      :testerItem="testerItem"
-      @refresh="fetchTester"
-    ></TesterItem>
+    <TesterItem></TesterItem>
   </div>
 </template>
 
@@ -16,31 +12,6 @@ import { testerInformation } from '@/api/auth';
 export default {
   name: 'TesterInfoPage',
   components: { TesterItem },
-  data() {
-    return {
-      testerItem: {
-        type: Object,
-        required: true,
-      },
-    };
-  },
-  methods: {
-    async fetchTester() {
-      try {
-        const userID = this.$store.state.UserID;
-        const { data } = await testerInformation(userID);
-        console.log(data);
-        this.testerItem = data;
-        return data;
-      } catch (error) {
-        console.log(error);
-      }
-    },
-  },
-
-  created() {
-    this.fetchTester();
-  },
 };
 </script>
